Eliminate compiler warnings (#1156)
* Fixed uninitialized variables in the GenRandom backend.

* Declared Eigen as a system library in order to silence warnings coming from Eigen include
files.

* Eliminated warnings from GNU compiler

* kick off tests again after oops change

* Added 'f' to float literal to avoid conversion from double to float.

---------

+++ src/engines/ioda/src/ioda/C/ioda_c_utils.cpp
@@ -187,7 +187,7 @@ void BlockingRead(int fd,void *buff,size_t sz)
 void Fwrite(const void *p,size_t osize,size_t cnt,FILE *fp)
 {
     int64_t e = fwrite(p,osize,cnt,fp);
-    if (e==cnt) return;
+    if (e==static_cast<int64_t>(cnt)) return;
     fprintf(stderr,"Fwrite failed ferror = %s\n",strerror(e));
     fatal_error();
 }
